docreader/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
bayoo-docx==0.2.17
black==21.5b1
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
colorama==0.4.4
comtypes==1.1.10
configparser==5.0.2
crayons==0.4.0
Django==3.1.7
django-colorfield==0.4.1
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-dotenv==1.4.2
django-filter==2.4.0
djangorestframework==3.12.4
docx==0.2.4
docx2pdf==0.1.7
google-api-core==1.26.3
google-api-python-client==2.4.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
importlib-metadata==1.7.0
isort==5.8.0
lxml==4.6.3
Markdown==3.3.4
mypy-extensions==0.4.3
oauthlib==3.1.0
packaging==20.9
pathspec==0.8.1
Pillow==8.2.0
protobuf==3.17.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2021.1
pywin32==227
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
selenium==3.141.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tqdm==4.60.0
typed-ast==1.4.3
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
webdriver-manager==3.4.2
zipp==3.4.1
